improve query speed by reducing output volume
